Add callout component in post-body
