+++ lib/api.js
@@ -36,6 +36,12 @@ content {
           code
           codeLanguage
         }
+        ... on Callout {
+          content {
+            json
+          }
+          title
+        }
       }
     }
     assets {
