+++ modules/blog/posts-list/posts-list.module.scss
@@ -27,7 +27,7 @@
 }
 
 .imageWrapper {
-  left: 4rem;
+  left: 17rem;
   opacity: 0.5;
   opacity: 0;
   position: absolute;
