+++ styles/_variables.scss
@@ -13,7 +13,9 @@ $color-grey-200: #d5d8da;
 $color-grey-500: #97a0a6;
 $color-grey-900: #4c5b64;
 $color-black-700: #242424;
+$color-white-100: #f9f9fa;
 $color-white-200: #f2f3f4;
+$color-white-500: #dfe2e4;
 $color-white-900: #c6cace;
 $color-slate: #2a3b46;
