+++ utils/blog/index.ts
@@ -36,3 +36,11 @@ export const getTiltEffectValues = (
 
   return { valueX, valueY };
 };
+
+export const dateFormatting = (date: string) => {
+  return new Date(date).toLocaleDateString('en-GB', {
+    day: 'numeric',
+    month: 'short',
+    year: 'numeric',
+  });
+};
