--- utils/date-formatting.ts
@@ -1,7 +1,0 @@
-export const dateFormatting = (date: string) => {
-  return new Date(date).toLocaleDateString('en-GB', {
-    day: 'numeric',
-    month: 'short',
-    year: 'numeric',
-  });
-};
